+++ velox/docs/configs.rst
@@ -959,7 +959,7 @@ Spark-specific Configuration
      - false
      - If true, Spark statistical aggregation functions including skewness, kurtosis will return NaN instead of NULL
        when dividing by zero during expression evaluation. Please note that Spark statistical aggregation functions
-       including stddev, stddev_samp, variance, var_samp, covar_samp and corr should be supported to respect this configuration.
+       including covar_samp and corr should be supported to respect this configuration.
 
 Tracing
 --------
