+++ velox/functions/sparksql/aggregates/tests/CMakeLists.txt
@@ -27,7 +27,8 @@ add_executable(
   MinMaxByAggregationTest.cpp
   ModeAggregateTest.cpp
   RegrReplacementAggregationTest.cpp
-  SumAggregationTest.cpp)
+  SumAggregationTest.cpp
+  VarianceAggregationTest.cpp)
 
 add_test(velox_functions_spark_aggregates_test
          velox_functions_spark_aggregates_test)
